refactor(home): extract the code of a banner view
